Remove unused agent deployer
